Wrong data type

+++ Game Commons/src/de/gymdon/inf1315/game/packet/PacketStartGame.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 
 public class PacketStartGame extends Packet {
 
-    public static final int ID = 3;
+    public static final short ID = 3;
     public String opponent;
     
     public PacketStartGame(Remote r) {
